pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vlsi-agent"
version = "0.1.0"
description = "Deterministic YAML-to-RTL FIFO starter flow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "jinja2>=3.1",
  "pyyaml>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
